pyproject.toml
[project]
name = "legal-eagle-mvp"
version = "0.1.0"
description = "Advanced AI Legal Intelligence Platform for Indian Law"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Your existing core dependencies (keeping the ones we need)
    "faiss-cpu>=1.7.4",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "python-docx>=1.1.0",
    "sentence-transformers>=2.2.2",
    "spacy>=3.8.7",
    "streamlit>=1.28.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    # CRITICAL: Replace google-genai with google-generativeai
    "google-generativeai>=0.3.2",
    # Essential Streamlit UI Components
    "streamlit-option-menu>=0.3.6",
    "streamlit-aggrid>=0.3.4",
    # Core Data Processing (MISSING)
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "polars>=0.19.0",
    "pyarrow>=13.0.0",
    "openpyxl>=3.1.0",
    # Security & Cryptography (MISSING - ESSENTIAL)
    "cryptography>=41.0.0",
    "bcrypt>=4.0.1",
    "pyjwt>=2.8.0",
    "passlib>=1.7.4",
    # HTTP & Async (MISSING)
    "requests>=2.31.0",
    "aiohttp>=3.8.5",
    "httpx>=0.24.1",
    "aiofiles>=23.2.1",
    # Date/Time Processing (MISSING)
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    # Configuration & Environment (MISSING)
    "python-dotenv>=1.0.0",
    "pydantic>=2.3.0",
    "pydantic-settings>=2.0.3",
    "toml>=0.10.2",
    # Logging (MISSING)
    "loguru>=0.7.0",
    "structlog>=23.1.0",
    # Caching & Performance (MISSING)
    "redis>=4.6.0",
    "diskcache>=5.6.3",
    # Text Processing Enhancement (MISSING)
    "nltk>=3.8.1",
    "textstat>=0.7.3",
    # Image Processing Enhancement (MISSING)
    "opencv-contrib-python>=4.8.0",
    "scikit-image>=0.21.0",
    # JSON & Serialization (MISSING)
    "orjson>=3.9.0",
    "ujson>=5.8.0",
    # UI Enhancement (MISSING)
    "tqdm>=4.65.0",
    "rich>=13.5.0",
    # System Monitoring (MISSING)
    "memory-profiler>=0.61.0",
    "psutil>=5.9.5",
    # ML Enhancement (MISSING)
    "datasets>=2.14.0",
    "accelerate>=0.23.0",
    # Database Support (Future-proofing)
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.7",
    "plotly>=5.15.0",
    "altair>=5.5.0",
    "google-cloud-documentai>=2.20.1",
    "google-cloud-aiplatform>=1.38.0",
    "google-auth>=2.22.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.10",
    "lxml>=4.9.3",
    "json5>=0.9.6",
    "textblob>=0.19.0",
    "hf-xet>=1.1.8",
    "nest-asyncio>=1.6.0",
    "google-oauth2-tool>=0.0.3",
    "pdfplumber>=0.9.0",
    "python-magic>=0.4.27",
    "asyncio-mqtt>=0.13.0",
    "jsonschema>=4.19.0",
    "pathlib2>=2.3.7",
    "streamlit-lottie>=0.0.5",
    "streamlit-elements>=0.1.0",
    "streamlit-card>=0.0.61",
    "wordcloud>=1.9.2",
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "sentry-sdk>=1.32.0",
    "prometheus-client>=0.18.0",
    "textacy>=0.13.0",
    "flashtext>=2.7",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.21.0",
    "pypdf2>=3.0.1",
    "st-annotated-text>=4.0.0",
]
[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0", 
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "jupyter>=1.1.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
